Create CityField component
